requirements.txt
streamlit
openai
python-dotenv
PyPDF2
azure-search-documents
azure-identity